requirements.txt
obd==0.7.1

opencv-python
flask

django==3.0.6
djangorestframework==3.11.0